+++ package/distance/polar.go
@@ -5,12 +5,12 @@ import (
 	"math/rand"
 )
 
-func PolarPoint(random *rand.Rand, r, a, b float64) Point {
-	randomRadius := r * random.Float64()
+func PolarPoint(random *rand.Rand, r, a, b float32) Point {
+	randomRadius := float64(r) * random.Float64()
 	randomAngle := random.Float64() * 2 * math.Pi
 
-	x := a + randomRadius*math.Cos(randomAngle)
-	y := b + randomRadius*math.Sin(randomAngle)
+	x := a + float32(randomRadius*math.Cos(randomAngle))
+	y := b + float32(randomRadius*math.Sin(randomAngle))
 
 	return Point{X: x, Y: y}
 }
